requirements.txt
contourpy==1.3.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
setuptools==76.0.0
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
